requirements.txt
Django==5.0.4
psycopg[binary]==3.1.18
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-environ==0.11.2
pytest==8.2.0
pytest-django==4.8.0
coverage==7.5.1
black==24.4.2
isort==5.13.2
flake8==7.0.0
gunicorn==22.0.0
dj-database-url==2.1.0 